Primera versión funcional (Muy simple)
Actualmente el programa hace lo siguiente:
1. Register funcional
2. Login funcional
3. Los mensajes se guardan en base de datos.
4. Los mensajes se muestran persistiendo en base de datos.

+++ src/main/java/org/sk/chattcp/functionality/threads/ThreadServerChat.java
@@ -42,7 +42,7 @@ public class ThreadServerChat extends Thread {
                     break;
                 }
                 String[] parts = cadena.split(":");
-                User sender = userRepository.findById(Long.parseLong(parts[0]));
+                User sender = userRepository.findById(Integer.parseInt(parts[0]));
                 String content = parts[1];
                 Message message = new Message();
                 message.setSender(sender);
